# faster subImage, a bit more