before:
`subImage ........................... 0.126 ms      0.139 ms    ±0.015  x1000`
after:
`subImage ........................... 0.024 ms      0.027 ms    ±0.003  x1000`